+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Added `tsconfig.tests.json` specifically for test configurations.
+
+### Changed
+
+- Consolidated `create-audio-frame-buffer-context.ts` into `audio-frame-buffer-context.ts` and removed `create-audio-frame-buffer-context.ts`.
+
 ## [0.1.0] - 2024-11-05
 
 ### Changed